issue132: make project more modular and fix routing
