+++ MultCo-ISD-API/MultCo-ISD-API/V1/Controllers/ServiceController.cs
@@ -532,413 +532,8 @@ namespace MultCo_ISD_API.V1.Controllers
 			return serviceDTO;
 		}
 		// ----------------- PUT lang ------------
-		[HttpPut("Language/{languageId}")]
-		[ProducesResponseType(204)]
-		[ProducesResponseType(400)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Writer")]
-#endif
-		public async Task<IActionResult> PutLanguage(int languageId, [FromBody] LanguageV1DTO languageDTO)
-		{
-			try
-			{
-				// Check to ensure service exists before calling contextmanager method.
-				var language = await _serviceContextManager.GetLanguageByIdAsync(languageId);
-				if (language == null)
-				{
-					return NotFound();
-				}
-				languageDTO.LanguageId = languageId;
-
-				await _serviceContextManager.PutLanguageAsync(languageDTO);
-				return NoContent();
-			}
-			catch (Exception e)
-			{
-				throw e;
-			}
-		}
-		[HttpPut("Community/{communityId}")]
-		[ProducesResponseType(204)]
-		[ProducesResponseType(400)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Writer")]
-#endif
-		public async Task<IActionResult> PutCommunity(int communityId, [FromBody] CommunityV1DTO communityDTO)
-		{
-			try
-			{
-				// Check to ensure service exists before calling contextmanager method.
-				var community = await _serviceContextManager.GetCommunityByIdAsync(communityId);
-				if (community == null)
-				{
-					return NotFound();
-				}
-				communityDTO.CommunityId = communityId;
-
-				await _serviceContextManager.PutCommunityAsync(communityDTO);
-				return NoContent();
-			}
-			catch (Exception e)
-			{
-				throw e;
-			}
-		}
-		[HttpPut("Division/{divisionId}")]
-		[ProducesResponseType(204)]
-		[ProducesResponseType(400)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Writer")]
-#endif
-		public async Task<IActionResult> PutDivision(int divisionId, [FromBody] DivisionV1DTO divisionDTO)
-		{
-			try
-			{
-				// Check to ensure service exists before calling contextmanager method.
-				var division = await _serviceContextManager.GetDivisionByIdAsync(divisionId);
-				if (division == null)
-				{
-					return NotFound();
-				}
-				divisionDTO.DivisionId = divisionId;
-
-				await _serviceContextManager.PutDivisionAsync(divisionDTO);
-				return NoContent();
-			}
-			catch (Exception e)
-			{
-				throw e;
-			}
-		}
-		[HttpPut("Department/{departmentId}")]
-		[ProducesResponseType(204)]
-		[ProducesResponseType(400)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Writer")]
-#endif
-		public async Task<IActionResult> PutDepartment(int departmentId, [FromBody] DepartmentV1DTO departmentDTO)
-		{
-			try
-			{
-				// Check to ensure service exists before calling contextmanager method.
-				var department = await _serviceContextManager.GetDepartmentByIdAsync(departmentId);
-				if (department == null)
-				{
-					return NotFound();
-				}
-				departmentDTO.DepartmentId = departmentId;
-
-				await _serviceContextManager.PutDepartmentAsync(departmentDTO);
-				return NoContent();
-			}
-			catch (Exception e)
-			{
-				throw e;
-			}
-		}
-		[HttpPut("Program/{programId}")]
-		[ProducesResponseType(204)]
-		[ProducesResponseType(400)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Writer")]
-#endif
-		public async Task<IActionResult> PutProgram(int programId, [FromBody] ProgramV1DTO programDTO)
-		{
-			try
-			{
-				// Check to ensure service exists before calling contextmanager method.
-				var program = await _serviceContextManager.GetProgramByIdAsync(programId);
-				if (program == null)
-				{
-					return NotFound();
-				}
-				programDTO.ProgramId = programId;
-
-				await _serviceContextManager.PutProgramAsync(programDTO);
-				return NoContent();
-			}
-			catch (Exception e)
-			{
-				throw e;
-			}
-		}
-
-		[HttpPut("Location/{locationId}")]
-		[ProducesResponseType(204)]
-		[ProducesResponseType(400)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Writer")]
-#endif
-		public async Task<IActionResult> PutLocation(int locationId, [FromBody] LocationV1DTO locationDTO)
-		{
-			try
-			{
-				// Check to ensure service exists before calling contextmanager method.
-				var location = await _serviceContextManager.GetLocationByIdAsync(locationId);
-				if (location == null)
-				{
-					return NotFound();
-				}
-				locationDTO.LocationId = locationId;
-
-				await _serviceContextManager.PutLocationAsync(locationDTO);
-				return NoContent();
-			}
-			catch (Exception e)
-			{
-				throw e;
-			}
-		}
-		[HttpPut("LocationType/{locationTypeId}")]
-		[ProducesResponseType(204)]
-		[ProducesResponseType(400)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Writer")]
-#endif
-		public async Task<IActionResult> PutLocationType(int locationTypeId, [FromBody] LocationTypeV1DTO locationTypeDTO)
-		{
-			try
-			{
-				// Check to ensure service exists before calling contextmanager method.
-				var locationType = await _serviceContextManager.GetLocationTypeByIdAsync(locationTypeId);
-				if (locationType == null)
-				{
-					return NotFound();
-				}
-				locationTypeDTO.LocationTypeId = locationTypeId;
-
-				await _serviceContextManager.PutLocationTypeAsync(locationTypeDTO);
-				return NoContent();
-			}
-			catch (Exception e)
-			{
-				throw e;
-			}
-		}
-		[HttpPut("Contact/{contactId}")]
-		[ProducesResponseType(204)]
-		[ProducesResponseType(400)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Writer")]
-#endif
-		public async Task<IActionResult> PutContact(int contactId, [FromBody] ContactV1DTO contactDTO)
-		{
-			try
-			{
-				// Check to ensure service exists before calling contextmanager method.
-				var contact = await _serviceContextManager.GetContactByIdAsync(contactId);
-				if (contact == null)
-				{
-					return NotFound();
-				}
-				contactDTO.ContactId = contactId;
-
-				await _serviceContextManager.PutContactAsync(contactDTO);
-				return NoContent();
-			}
-			catch (Exception e)
-			{
-				throw e;
-			}
-		}
 		// ----------------- end put -------------------
-		// ----------------- start get -------------------
-		/// <summary>
-		/// Get language by id 
-		/// </summary>
-		/// <remarks>
-		/// Returns language with a matching language id
-		/// </remarks>
-		/// <returns></returns>
-		[SwaggerOperation(Tags = new[] { "Reader" })]
-		//GET: api/Services/lang?=languageId
-		[HttpGet]
-		[Route("[action]")]
-		[ProducesResponseType(typeof(ServiceV1DTO), (int)HttpStatusCode.OK)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Reader")]
-#endif
-		public async Task<IActionResult> GetLanguage([FromQuery][Required] int languageId)
-		{
-			var languageDTO = await _serviceContextManager.GetLanguageByIdAsync(languageId);
-
-			if (languageDTO == null)
-			{
-				return NotFound("No languages from given id found.");
-			}
-
-			return Ok(languageDTO);
-		}
-		/// <summary>
-		/// Get community by id 
-		/// </summary>
-		/// <remarks>
-		/// Returns community with a matching language id
-		/// </remarks>
-		/// <returns></returns>
-		[SwaggerOperation(Tags = new[] { "Reader" })]
-		[HttpGet]
-		[Route("[action]")]
-		[ProducesResponseType(typeof(ServiceV1DTO), (int)HttpStatusCode.OK)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Reader")]
-#endif
-		public async Task<IActionResult> GetCommunity([FromQuery][Required] int communityId)
-		{
-			var communityDTO = await _serviceContextManager.GetCommunityByIdAsync(communityId);
-
-			if (communityDTO == null)
-			{
-				return NotFound("No community from given id found.");
-			}
-
-			return Ok(communityDTO);
-		}
-
-		/// <summary>
-		/// Get department by id 
-		/// </summary>
-		/// <remarks>
-		/// Returns department with a matching department id
-		/// </remarks>
-		/// <returns></returns>
-		[SwaggerOperation(Tags = new[] { "Reader" })]
-		[HttpGet]
-		[Route("[action]")]
-		[ProducesResponseType(typeof(ServiceV1DTO), (int)HttpStatusCode.OK)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Reader")]
-#endif
-		public async Task<IActionResult> GetDepartment([FromQuery][Required] int departmentId)
-		{
-			var departmentDTO = await _serviceContextManager.GetDepartmentByIdAsync(departmentId);
-
-			if (departmentDTO == null)
-			{
-				return NotFound("No department from given id found.");
-			}
-
-			return Ok(departmentDTO);
-		}
-
-		/// <summary>
-		/// Get division by id 
-		/// </summary>
-		/// <remarks>
-		/// Returns division with a matching division id
-		/// </remarks>
-		/// <returns></returns>
-		[SwaggerOperation(Tags = new[] { "Reader" })]
-		[HttpGet]
-		[Route("[action]")]
-		[ProducesResponseType(typeof(ServiceV1DTO), (int)HttpStatusCode.OK)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Reader")]
-#endif
-		public async Task<IActionResult> GetDivision([FromQuery][Required] int divisiontId)
-		{
-			var divisiontDTO = await _serviceContextManager.GetDivisionByIdAsync(divisiontId);
-
-			if (divisiontDTO == null)
-			{
-				return NotFound("No division from given id found.");
-			}
-
-			return Ok(divisiontDTO);
-		}
-
-		/// <summary>
-		/// Get program by id 
-		/// </summary>
-		/// <remarks>
-		/// Returns program with a matching program id
-		/// </remarks>
-		/// <returns></returns>
-		[SwaggerOperation(Tags = new[] { "Reader" })]
-		[HttpGet]
-		[Route("[action]")]
-		[ProducesResponseType(typeof(ServiceV1DTO), (int)HttpStatusCode.OK)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Reader")]
-#endif
-		public async Task<IActionResult> GetProgram([FromQuery][Required] int programtId)
-		{
-			var programDTO = await _serviceContextManager.GetProgramByIdAsync(programtId);
-
-			if (programDTO == null)
-			{
-				return NotFound("No program from given id found.");
-			}
-
-			return Ok(programDTO);
-		}
-
-		/// <summary>
-		/// Get location by id 
-		/// </summary>
-		/// <remarks>
-		/// Returns location with a matching location id
-		/// </remarks>
-		/// <returns></returns>
-		[SwaggerOperation(Tags = new[] { "Reader" })]
-		[HttpGet]
-		[Route("[action]")]
-		[ProducesResponseType(typeof(ServiceV1DTO), (int)HttpStatusCode.OK)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Reader")]
-#endif
-		public async Task<IActionResult> GetLocation([FromQuery][Required] int locationtId)
-		{
-			var locationDTO = await _serviceContextManager.GetLocationByIdAsync(locationtId);
-
-			if (locationDTO == null)
-			{
-				return NotFound("No location from given id found.");
-			}
-
-			return Ok(locationDTO);
-		}
-
-		/// <summary>
-		/// Get locationType by id 
-		/// </summary>
-		/// <remarks>
-		/// Returns locationType with a matching locationType id
-		/// </remarks>
-		/// <returns></returns>
-		[SwaggerOperation(Tags = new[] { "Reader" })]
-		[HttpGet]
-		[Route("[action]")]
-		[ProducesResponseType(typeof(ServiceV1DTO), (int)HttpStatusCode.OK)]
-		[ProducesResponseType(404)]
-#if AUTH
-        [Authorize(Policy = "Reader")]
-#endif
-		public async Task<IActionResult> GetLocationType([FromQuery][Required] int locationTypetId)
-		{
-			var locationTypeDTO = await _serviceContextManager.GetLocationTypeByIdAsync(locationTypetId);
-
-			if (locationTypeDTO == null)
-			{
-				return NotFound("No locationType from given id found.");
-			}
-
-			return Ok(locationTypeDTO);
-		}
+		// ----------------- start get -----------------
 		// ----------------- end get -------------------
 	}
 }
